Caps lock behaves correctly
Fixed caps lock functionality. Also forced caps lock off
when printing out the version string.
