+++ src/firmware/lib/adbuino/src/usbinterface.cpp
@@ -24,14 +24,7 @@
 
 void UsbInterface::Init(){
     Usb = new USB();
-    // Hub1 = new USBHub(Usb);
-    // Hub2 = new USBHub(Usb);
-    // Hub3 = new USBHub(Usb);
-    // Hub4 = new USBHub(Usb);
-    // HidKeyboard = new HIDBoot<USB_HID_PROTOCOL_KEYBOARD>(Usb);
-    // HidMouse = new HIDBoot<USB_HID_PROTOCOL_MOUSE>(Usb);
-    // MousePrs = new MouseRptParser();
-    // KeyboardPrs = new KbdRptParser();
+
 
     if (Usb->Init() == -1)
         Serial.println("OSC did not start.");
