+++ src/firmware/lib/usb/include/usbkbdparser.h
@@ -25,34 +25,6 @@
 
 #include "platformkbdparser.h"
 
-#include "scqueue.h"
-
-using simple_circular_queue::SCQueue;
-
-#define KEYBOARD_QUEUE_CAPACITY (20)
-
-class KeyEvent
-{
-public:
-    static const uint8_t NoKey = 0xFF;
-    static const uint8_t KeyDown = 0x01;
-    static const uint8_t KeyUp = 0x02;
-    inline uint8_t GetKeycode() { return m_keycode; }
-    inline bool IsKeyUp() { return m_key_updown == KeyUp; }
-    inline bool IsKeyDown() { return m_key_updown == KeyDown; }
-    KeyEvent(uint8_t KeyCode, uint8_t KeyUpDown, uint8_t mod)
-    {
-        m_key_updown = KeyUpDown;
-        m_keycode = KeyCode;
-        m_mod = mod;
-    }
-
-protected:
-    uint8_t m_keycode;
-    uint8_t m_key_updown;
-    uint8_t m_mod;
-};
-
 class KbdRptParser : public PlatformKbdParser
 {
 public:
@@ -93,6 +65,6 @@ protected:
     static const int Led2CapsLockFlag = 6;
     static const int Led1NumLockFlag = 7;
 
-    SCQueue<KeyEvent*, KEYBOARD_QUEUE_CAPACITY> m_keyboard_events;
+    
 
 };
